+++ SemiStep/Shared/Registries/IActionRegistry.cs
@@ -7,7 +7,7 @@ public interface IActionRegistry
 	void Initialize(IReadOnlyDictionary<int, ActionDefinition> actions);
 	ActionDefinition GetAction(int id);
 	ActionDefinition GetActionByName(string name);
-	bool ActionExists(short id);
+	bool ActionExists(int id);
 	bool ActionExistsByName(string name);
 	IReadOnlyList<ActionDefinition> GetAll();
 }
